fix/translate: dialogue partially works and in English
- The NPC talks again, and only in English.
- All dialogue is translated, but not all is used.
- Steps, Subtasks, skills, popups, are all translated.
- Fish sound is restored.
- Tablet works with some minor problems.

Refs: #612, #599, #613
