+++ Assets/FishMaintenance/Boss.asset
@@ -24,6 +24,8 @@ MonoBehaviour:
   ShouldFollow: 0
   DialogueTreesSO:
   - {fileID: 11400000, guid: 48653bbc2b4e74648abc1e3560964e74, type: 2}
+  - {fileID: 11400000, guid: 022a82fdecd041e4dba991b8a0575791, type: 2}
+  - {fileID: 11400000, guid: e4b326c65881be646960b68a5940cc68, type: 2}
   DialogueTreeJSON: []
   runtimeAnimatorController: {fileID: 9100000, guid: acb4b9981fa45ce46adb4867634b31ab,
     type: 2}
