--- Assets/FishMaintenance/PauseSjef.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 022a82fdecd041e4dba991b8a0575791
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
