+++ Assets/FishMaintenance/Scripts/DailyCheckStep.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7bcb64a152bf7c845a5cbe11d99f46e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
