+++ Assets/FishMaintenance/Tasks/Subtasks/Deadfish.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 694b8dd6ca4134b4d82fc6b2dc8d2848, type: 3}
   m_Name: Deadfish
   m_EditorClassIdentifier: 
-  _subtaskName: "D\xF8dfisk h\xE5ndtering"
+  _subtaskName: Handling of dead fish
   _description: TBA
   _autocompleate: 1
   _stepList:
